create elements for front and back covers
