+++ src/styles/card.module.css
@@ -73,9 +73,3 @@
 	display: flex;
 	flex-wrap: wrap;
 }
-
-.infoIcon {
-	padding-right: 5px;
-}
-
-
